% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotICC}
\alias{plotICC}
\title{Plot item characteristic curves (ICCs)}
\usage{
plotICC(IP, Th, items = NULL)
}
\arguments{
\item{IP}{Object of class \code{GGUM}.}

\item{Th}{Theta estimates from function \code{Theta.EAP()}.}

\item{items}{Vector indicating the items for which the ICCs are to be 
plotted. Default is all items.}
}
\value{
The function returns the correlation between observed and expected 
  item scores (missing values pairwise removed).
}
\description{
\code{plot.ICC} plots the ICCs for the GUM and the GGUM.
}
\section{Details}{
 This function plots the item characteristic curves (ICCs).
}

\examples{
\dontrun{
# Generate data:
set.seed(1); C <- sample(3:5, 10, replace = TRUE)
gen <- GenData.GGUM(2000, 10, C, seed = 125)
# Fit the GGUM:
fit <- GGUM(gen$data, C)
th  <- Theta.EAP(fit)
# Plot the ICCs:
plotICC(fit, th)
}

}
\author{
Jorge N. Tendeiro, \email{j.n.tendeiro@rug.nl}
}
