% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotIIF}
\alias{plotIIF}
\title{Plot item information functions (IIFs)}
\usage{
plotIIF(IP, Th, items = NULL)
}
\arguments{
\item{IP}{Object of class \code{GGUM}.}

\item{Th}{Theta estimates from function \code{Theta.EAP()}.}

\item{items}{Vector indicating the items for which the ICCs are to be 
plotted. Default is all items.}
}
\value{
The function returns the (x, y) coordinates of the IIFs.
}
\description{
\code{plot.IIF} plots the IIFs for the GUM and the GGUM.
}
\section{Details}{
 This function plots the item information functions (IIFs).
}

\examples{
\dontrun{
# Generate data:
set.seed(1); C <- sample(3:5, 10, replace = TRUE)
gen <- GenData.GGUM(2000, 10, C, seed = 125)
# Fit the GGUM:
fit <- GGUM(gen$data, C)
th  <- Theta.EAP(fit)
# Plot the IIFs:
plot.IIF(fit, th)
}

}
\author{
Jorge N. Tendeiro, \email{j.n.tendeiro@rug.nl}
}
