% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotTIF}
\alias{plotTIF}
\title{Plot test information function (TIF)}
\usage{
plotTIF(IP, Th)
}
\arguments{
\item{IP}{Object of class \code{GGUM}.}

\item{Th}{Theta estimates from function \code{Theta.EAP()}.}
}
\value{
The function returns the (x, y) coordinates of the TIF.
}
\description{
\code{plot.TIF} plots the TIF for the GUM and the GGUM.
}
\section{Details}{
 This function plots the test information function (TIF).
}

\examples{
\dontrun{
# Generate data:
set.seed(1); C <- sample(3:5, 10, replace = TRUE)
gen <- GenData.GGUM(2000, 10, C, seed = 125)
# Fit the GGUM:
fit <- GGUM(gen$data, C)
th  <- Theta.EAP(fit)
# Plot the TIF:
plot.TIF(fit, th)
}

}
\author{
Jorge N. Tendeiro, \email{j.n.tendeiro@rug.nl}
}
