% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGUM2004.R
\name{read.item.GGUM2004}
\alias{read.item.GGUM2004}
\title{Read GGUM2004 Item Estimates}
\usage{
read.item.GGUM2004(I, C, model = "GGUM",
  tempfolder = "C:/GGUM2004/TEMPFILE")
}
\arguments{
\item{I}{the number of items.}

\item{C}{either a number or a vector. C is the number of observable response
categories minus 1.}

\item{model}{A string identifying the model. Possible values are "GUM" or 
"GGUM" (default).}

\item{tempfolder}{the path where GGUM2004 save its output.By default it is
"C:/GGUM2004/TEMPFILE".}
}
\value{
\code{read.item.GGUM2004} returns a list cointaning the following
  components: \itemize{ \item \code{delta} a vector with delta estimates
  \item \code{alpha} a vector with alpha estimates \item \code{taus} a matrix
  with taus estimates \item \code{deltaSE} a vector with delta standard
  errors \item \code{alphaSE} a vector with alpha standard errors \item
  \code{tausSE} a matrix with taus standard errors }
}
\description{
\code{read.item.GGUM2004} reads the item parameter file from
  GGUM2004. It reads the delta estimates, the alpha estimates, the taus
  estimates, and their standard errors.
}
