% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGUM2004.R
\name{read.person.GGUM2004}
\alias{read.person.GGUM2004}
\title{Read GGUM2004 Person Estimates}
\usage{
read.person.GGUM2004(N, tempfolder = "C:/GGUM2004/TEMPFILE")
}
\arguments{
\item{N}{the number of persons to be read.}

\item{tempfolder}{the path where GGUM2004 save its output.By default it is
"C:/GGUM2004/TEMPFILE".}
}
\value{
A matrix (dim N*2) is returned with the estimates in the first column
  and their standard errors in the second column.
}
\description{
\code{read.person.GGUM2004} reads the person parameter file from
  GGUM2004. It reads the estimates and the standard errors.
}
