% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGUM2004.R
\name{export.GGUM2004}
\alias{export.GGUM2004}
\title{Exports Data to GGUM2004 Format}
\usage{
export.GGUM2004(data, file.name = "MyData")
}
\arguments{
\item{data}{a matrix with the data.}

\item{file.name}{a character string which give the name to the data file.}
}
\description{
\code{export.GGUM2004} writes the data to a text file according
  to the format required by GGUM2004.
}
