% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGUM2004.R
\name{run.GGUM2004}
\alias{run.GGUM2004}
\title{Runs GGUM2004}
\usage{
run.GGUM2004(cmd.file, I, C, N, SE = TRUE, model = "GGUM",
  cmd.dir = "C:/GGUM2004")
}
\arguments{
\item{cmd.file}{the name of the GGUM2004 command file.}

\item{I}{the number of items.}

\item{C}{either a number or a vector. C is the number of observable response
categories minus 1.}

\item{N}{the number of persons to be read.}

\item{SE}{?????????????????????????}

\item{model}{A string identifying the model. Possible values are "GUM" or 
"GGUM" (default).}

\item{cmd.dir}{the directory of GGUM2004 program. It is predefined as
"C:/GGUM2004"}
}
\value{
\code{run.GGUM2004} returns a list cointaning the following
  components: \itemize{ \item \code{time} the spent time in the analysis
  \item \code{delta} a vector with delta estimates \item \code{alpha} a
  vector with alpha estimates \item \code{taus} a matrix with taus estimates
  \item \code{theta} a vector with theta estimates}
}
\description{
\code{run.GGUM2004} sends the command file to GGUM2004 and runs
  it. It returns the time and the parameter estimates.
}
