% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGUM2004.R
\name{write.GGUM2004}
\alias{write.GGUM2004}
\title{Writes a Generic Command File for GGUM2004}
\usage{
write.GGUM2004(cmd.file, data.file, I, C, cutoff = 2, model = "GGUM",
  cmd.dir = "C:/GGUM2004")
}
\arguments{
\item{cmd.file}{A character string which give a name to the command file.}

\item{data.file}{A character string with the complete path of the data file.}

\item{I}{The number of items.}

\item{C}{Either a number or a vector of I elements. C is the number of
observable response categories minus 1.}

\item{cutoff}{either a number or a vectorof I elements which defines the
cutoff value.}

\item{model}{A string identifying the model. Possible values are "GUM" or 
"GGUM" (default).}

\item{cmd.dir}{the directory for GGUM2004 program. It is predefined as 
"C:/GGUM2004".}
}
\description{
\code{write.GGUM2004} allows to partially customise the GGUM2004
  command file according to the test characteristics. The file is 
  automatically saved in the GGUM2004 predefined installation folder.
}
